+++ lib/airbrake-ruby/notifier.rb
@@ -101,7 +101,17 @@ module Airbrake
 
     # @macro see_public_api_method
     def notify_request(request_info)
-      @route_sender.notify_request(request_info)
+      promise = Airbrake::Promise.new
+
+      if @config.ignored_environment?
+        return promise.reject("The '#{@config.environment}' environment is ignored")
+      end
+
+      unless @config.route_stats
+        return promise.reject("The Route Stats feature is disabled")
+      end
+
+      @route_sender.notify_request(request_info, promise)
     end
 
     # @return [String] customized inspect to lessen the amount of clutter